test: 뷰 컨트롤러 관련 기능 테스트 코드 통합
