--- src/test/java/cart/SettingsIntegrationTest.java
@@ -1,33 +1,0 @@
-package cart;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class SettingsIntegrationTest {
-
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @Test
-    @DisplayName("")
-    void settingsTest() {
-        RestAssured.given()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .get("/settings")
-                .then()
-                .statusCode(HttpStatus.OK.value());
-    }
-}
